feat(go): update installation logic for Arch

+++ go/init.sh
@@ -1,13 +1,21 @@
 #!/usr/bin/bash
 
 echo '*****************************************'
-echo '*             Installing go             *'
+echo '*             Installing Go             *'
 echo '*****************************************'
 
-file=$(curl -L -s 'https://go.dev/dl' | grep -m 1 '/dl/go.*.linux-amd64' | awk -F "/" '{print $3}' | awk -F '">' '{print $1}')
+if [[ $INSTALL_OS = 'arch' ]]; then
+  sudo pacman -S --noconfirm --needed go
+fi
 
-curl -L "https://go.dev/dl/$file" -o go.tar.gz
-sudo rm -rf /usr/local/go
-sudo tar -C /usr/local -xzf go.tar.gz
+if [[ $INSTALL_OS = 'popos' ]]; then
+  file=$(curl -L -s 'https://go.dev/dl' | grep -m 1 '/dl/go.*.linux-amd64' | awk -F "/" '{print $3}' | awk -F '">' '{print $1}')
 
-rm go.tar.gz
+  mkdir -p /tmp/dotfiles
+
+  curl -L "https://go.dev/dl/$file" -o /tmp/dotfiles/go.tar.gz
+  sudo rm -rf /usr/local/go
+  sudo tar -C /usr/local -xzf /tmp/dotfiles/go.tar.gz
+
+  rm -rf /tmp/dotfiles
+fi
